move stuff from main to syncthing with minimal mod
