+++ lib/syncthing/cpuusage_windows.go
@@ -6,7 +6,7 @@
 
 //+build windows
 
-package main
+package syncthing
 
 import "syscall"
 import "time"
